'driver': Don't highlight in isearch
zsh version 5.2 and lower don't support ISEARCHMATCH_ACTIE and
we are unable to re-apply zle_highlight on top. Therefore it is
impossible to see the underlined matched area.

Since that information is more important, completely disable
highlighting in isearch in that case.

To do that, we need to make sure we are actually called when
something changes in isearch.

Trumps #257.